requirements.txt
datasets==2.21.0
fpdf==1.7.2
openai==1.56.0
langchain==0.3.9
langchain-community==0.3.9
langchain-openai==0.2.10
langchain-experimental==0.3.3
langchain-neo4j==0.1.1
neo4j==5.27.0
python-dotenv==1.0.1
pypdf==4.3.1
pydantic==2.9.1
pydantic_core==2.23.3
streamlit==1.35.0
flask